feat: allow lobby reentry (#47)
This PR adds shutdown of the current network runner when the game
restarts. This will end the issue of leaked network runners causing the
Lobby to stay open. If, in the future, we need to accommodate multiple
scenes we can move the `NetworkManager` into its own prefab and make it
`DontDestroyOnLoad`, but at the moment I don't there is any benefit to
doing this.